+++ lib/vagrant/errors.rb
@@ -203,6 +203,10 @@ module Vagrant
       error_key(:home_dir_not_accessible)
     end
 
+    class ForwardPortAdapterNotFound < VagrantError
+      error_key(:forward_port_adapter_not_found)
+    end
+
     class ForwardPortAutolistEmpty < VagrantError
       error_key(:auto_empty, "vagrant.actions.vm.forward_ports")
     end